+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+- Add the enabled status and receiver to the Laravel about command if possible.
+
+
 ## 4.1.0 - 2022-12-20
 
 ## Added